requirements.txt
pandas
babel
mysql-connector-python==8.0.33
openpyxl